+++ src/Sulu/Bundle/MediaBundle/Media/ImageConverter/Transformation/SharpenTransformation.php
@@ -1,0 +1,30 @@
+<?php
+
+/*
+ * This file is part of Sulu.
+ *
+ * (c) MASSIVE ART WebServices GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace Sulu\Bundle\MediaBundle\Media\ImageConverter\Transformation;
+
+use Imagine\Image\ImageInterface;
+
+/**
+ * Add the sharpen effect to an image.
+ */
+class SharpenTransformation implements TransformationInterface
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function execute(ImageInterface $image, $parameters)
+    {
+        $image->effects()->sharpen();
+
+        return $image;
+    }
+}