+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 CHANGELOG for Sulu
 ==================
 * dev-master
+    * FEATURE     #3028 [MediaBundle]             Added blur, grayscale, gamma, negative and sharpen transformation to media
     * HOTFIX      #3750 [PreviewBundle]           Fixed refresh preview
 
 * 1.6.14 (2018-02-06)